accident detection et alerte automatique/requirement.txt
streamlit
Pillow
numpy
joblib
fastai
fastai 
fastbook 
torchvision 
timm 
opencv-python
